# Node48 add: unroll nullptr-searching loop once